+++ agente/src/frsf/cidisi/exercise/diagrama/search/actions/IrNorte.java
@@ -70,7 +70,7 @@ public class IrNorte extends SearchAction {
 	 */
 	@Override
 	public Double getCost() {
-		return new Double(0);
+		return new Double(1);
 	}
 
 	/**
