+++ agente/src/frsf/cidisi/exercise/diagrama/search/actions/IrOeste.java
@@ -67,7 +67,7 @@ public class IrOeste extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(1);
     }
 
     /**
